Extract mid only mode logic to separate file
